pyproject.toml
[project]
name = "conversatioagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "dotenv>=0.9.9",
    "ipython>=9.9.0",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "langchain>=1.2.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.6",
    "openai>=2.15.0",
    "pypdf>=6.6.0",
    "qdrant-client>=1.16.2",
    "unstructured>=0.18.27",
]
